move CheckoutProvider, useCheckout, CurrencySelectorElement
