+++ package.json
@@ -8,6 +8,20 @@
   "browser:min": "dist/react-stripe.umd.min.js",
   "browser": "dist/react-stripe.umd.js",
   "types": "dist/react-stripe.d.ts",
+  "exports": {
+    ".": {
+      "require": "./dist/react-stripe.js",
+      "import": "./dist/react-stripe.esm.mjs",
+      "default": "./dist/react-stripe.esm.mjs",
+      "types": "./dist/react-stripe.d.ts"
+    },
+    "./checkout": {
+      "require": "./dist/checkout.js",
+      "import": "./dist/checkout.esm.mjs",
+      "default": "./dist/checkout.esm.mjs",
+      "types": "./dist/checkout.d.ts"
+    }
+  },
   "scripts": {
     "test": "yarn run lint && yarn run lint:prettier && yarn run test:unit && yarn test:package-types && yarn run typecheck",
     "test:package-types": "attw --pack .",
