+++ src/checkout/components/CheckoutProvider.test.tsx
@@ -3,10 +3,10 @@ import {render, act, waitFor} from '@testing-library/react';
 import {renderHook} from '@testing-library/react-hooks';
 
 import {CheckoutProvider, useCheckout} from './CheckoutProvider';
-import {Elements} from './Elements';
-import {useStripe} from './useStripe';
-import * as mocks from '../../test/mocks';
-import makeDeferred from '../../test/makeDeferred';
+import {Elements} from '../../components/Elements';
+import {useStripe} from '../../components/useStripe';
+import * as mocks from '../../../test/mocks';
+import makeDeferred from '../../../test/makeDeferred';
 
 describe('CheckoutProvider', () => {
   let mockStripe: any;
